Shell

+++ i-rex.sh
@@ -2,8 +2,15 @@
 
 if [ $# -eq 0 ] 
 then
-    echo "Usage: " $0 " <path-to-index>";
+    echo "Usage: i-rex <path-to-index>";
     exit 1;
 fi
 
-java -cp dist/i-rex.jar irex.IRex $1
+irexJarPath=dist/i-rex.jar
+if test -f "$irexJarPath";
+then
+    java -cp dist/i-rex.jar irex.IRex $1
+else
+    java -cp $HOME/bin/dist/i-rex.jar irex.IRex $1
+fi
+
